requirements.txt
py-desmume